Restructure some tests
